Changed file format to use linux return codes instead of dos for all files.
Added SMARTCTL_POWERMODE environment variable:
This is used for setting the nocheck option for smartctl. Setting this to sleep or standby will prevent the system from spinning up the harddrives for the smart check.
Valid values are: never,sleep,standby,idle.
The default is set to: never

Changed the docker-compose.yml and added the new SMARTCTL_POWERMODE environment variable to standby

Changed build.yml to dutchkarma repo